Implemented a 'simple' example server.
The simplest possible implementation of a Sack compliant server is one
that simply listens on a socket and accepts connections as they are
'wait'ed for. This commit implements that simple server as a baseline.
In future, specs and such will be built around this server as the
'reference' sack server.