Made the behaviour of a closed connection well defined.
When a request that has been closed is written to, it will error with
Sack::ClosedError indicating what connection was closed. This allows
closed client connections to be cleaned up in a timely manner.

Also implemented a connection list class that maintains a set of
connections and removes them lazily when they're closed. Its #each
method will even detect a closed connection error and remove it from the
list for that reason.